Break out cmd/yggdrasil a bit, tweaks
